feat(agentic-org): Phase 5 — bootable process host + container substrate + real CockroachDB-dialect fixes
Process host (apps/workers/src/main.ts):
- runMain(overrides?) reads env -> parseWorkerRuntimeConfigFromEnv -> builds
  pg/nats/telemetry -> migration bootstrap + readiness -> composes durable
  ports -> worker runtime -> process loop. Real SIGINT/SIGTERM signal source +
  setTimeout sleeper. Config error -> stderr -> exit 1 (graceful, never crash).
  import.meta.url main-module guard. node-process-host.d.ts ambient typing.

Container substrate:
- Dockerfile: node:24-slim, npm ci --omit=dev, USER node,
  ENTRYPOINT node --experimental-strip-types apps/workers/src/main.ts.
  Verified: exits 1 with "missing_cockroach_database_url" (graceful abort).
- docker-compose.yml: cockroach v26.2.1 (single-node --insecure, healthcheck),
  nats 2.14.1-alpine (-js), worker. .dockerignore excludes node_modules/tests/docs.

Real CockroachDB-dialect fixes (only visible against a live cluster):
- sql-statement-splitter.ts: splitSqlStatements() respects single-quoted
  strings ('' escape) + line comments. CockroachDB runs a multi-statement
  string as ONE implicit txn and forbids referencing a column added by an
  earlier DDL stmt in the same txn (ADD COLUMN version; UPDATE ... version ->
  SQLSTATE 42703). Migration runner + bootstrapper now execute per-statement.
- reaction-plan-work-queue: $N::INT8 cast before * INTERVAL '1 ms'
  (interval * interval is unsupported; param defaults to interval).

Tests: 490 total, 486 pass, 0 fail, 4 skipped. tsc 0 errors.
Golden-SQL assertions re-derived from splitSqlStatements (same source of truth).
All 4 live cockroach-integration tests pass against real Cockroach + NATS.

+++ agentic-organization/packages/state-cockroach/src/cockroach-reaction-plan-work-queue.ts
@@ -289,7 +289,7 @@ const CockroachReactionPlanWorkQueueSql = {
       status = '${ReactionPlanStatus.Claimed}',
       claim_id = $2,
       claimed_at = now(),
-      claim_expires_at = now() + ($3 * INTERVAL '1 millisecond')
+      claim_expires_at = now() + ($3::INT8 * INTERVAL '1 millisecond')
     WHERE reaction_plan_id IN (
       SELECT reaction_plan_id
       FROM ${CockroachTableName.ReactionPlans}
@@ -348,7 +348,7 @@ const CockroachReactionPlanWorkQueueSql = {
       claim_expires_at = CASE WHEN ($3->>'retryable')::BOOL AND attempt_count + 1 < $5 THEN NULL ELSE claim_expires_at END,
       attempt_count = attempt_count + 1,
       next_attempt_at = CASE
-        WHEN ($3->>'retryable')::BOOL AND attempt_count + 1 < $5 THEN now() + ($4 * INTERVAL '1 millisecond')
+        WHEN ($3->>'retryable')::BOOL AND attempt_count + 1 < $5 THEN now() + ($4::INT8 * INTERVAL '1 millisecond')
         ELSE NULL
       END,
       failed_at = now(),
